+++ tests/testthat/test-get_ifdata_registry.R
@@ -1,4 +1,4 @@
 # data <- get_ifdata_registry(
-#   year = 2024,
-#   month = 12
+#   year = c(2023, 2024),
+#   month = c(12,6)
 # )